Refactor: Break progress chart into components
